+++ src/main/java/com/unstoppabledomains/exceptions/ns/NSExceptionCode.java
@@ -13,5 +13,5 @@ public enum NSExceptionCode {
   UnsupportedCurrency,
   NotImplemented,
   InconsistentDomainArray,
-  MaxThreadLimit;
+  InvalidDomain;
 }